refactor(organon): extract energeia::qa submodule (#2642)
Move dokimasia and diorthosis tool definitions and executors to qa.rs.
